Move GenericType parser out of core
