+++ packages/walt-compiler/src/core/types.js
@@ -1,41 +1,6 @@
-import curry from 'curry';
 import Syntax from 'walt-syntax';
 import { mapNode } from 'walt-parser-tools/map-node';
 import walkNode from 'walt-parser-tools/walk-node';
-
-const mapGeneric = curry((options, node, _) => {
-  const { types } = options;
-  const [generic] = node.params;
-  const [T] = generic.params;
-  const realType = types[T.value];
-  const [args, result] = realType.params;
-  // Patch the node to be a real type which we can reference later
-  const patch = {
-    ...realType,
-    range: generic.range,
-    value: node.value,
-    meta: { ...realType.meta, CLOSURE_TYPE: generic.value === 'Lambda' },
-    params: [
-      {
-        ...args,
-        params: [
-          {
-            ...args,
-            params: [],
-            type: 'i32',
-            value: 'i32',
-            Type: Syntax.Type,
-          },
-          ...args.params,
-        ],
-      },
-      result,
-    ],
-  };
-  types[patch.value] = patch;
-
-  return patch;
-});
 
 export default function typePlugin() {
   return {
@@ -90,7 +55,6 @@ export default function typePlugin() {
               types[node.value] = parsed;
               return parsed;
             },
-            [Syntax.GenericType]: mapGeneric({ types }),
           })(ast);
 
           return next([astWithTypes, context]);
